+++ src/Contract.php
@@ -30,6 +30,7 @@ use OpenWeb3\Validators\StringValidator;
 use OpenWeb3\Validators\TagValidator;
 use OpenWeb3\Validators\QuantityValidator;
 use OpenWeb3\Formatters\AddressFormatter;
+use RuntimeException;
 use stdClass;
 
 
@@ -595,9 +596,9 @@ class Contract
 
     /**
      * getData
-     * Get the function method call data.
-     * With this function, you can send signed contract function transaction.
-     * 1. Get the funtion data with params.
+     * Get the contract method's call data.
+     * With this function, you can send signed contract method transactions.
+     * 1. Get the method data with parameters.
      * 2. Sign the data with user private key.
      * 3. Call sendRawTransaction.
      * 
@@ -663,5 +664,95 @@ class Contract
             return $functionData;
         }
         return '';
+    }
+
+    /**
+     * getEventLogs
+     *
+     * @param string $eventName
+     * @param string|int $fromBlock
+     * @param string|int $toBlock
+     * @return array
+     */
+    public function getEventLogs(string $eventName, $fromBlock = 'latest', $toBlock = 'latest')
+    {
+        //try to ensure block numbers are valid together
+        if ($fromBlock !== 'latest') {
+            if (!is_int($fromBlock) || $fromBlock < 1) {
+                throw new InvalidArgumentException('Please make sure fromBlock is a valid block number');
+            } else if ($toBlock !== 'latest' && $fromBlock > $toBlock) {
+                throw new InvalidArgumentException('Please make sure fromBlock is equal or less than toBlock');
+            }
+        }
+
+        if ($toBlock !== 'latest') {
+            if (!is_int($toBlock) || $toBlock < 1) {
+                throw new InvalidArgumentException('Please make sure toBlock is a valid block number');
+            } else if ($fromBlock === 'latest') {
+                throw new InvalidArgumentException('Please make sure toBlock is equal or greater than fromBlock');
+            }
+        }
+
+        $eventLogData = [];
+
+        //ensure the event actually exists before trying to filter for it
+        if (!array_key_exists($eventName, $this->events)) {
+            throw new InvalidArgumentException("'{$eventName}' does not exist in the ABI for this contract");
+        }
+
+        //indexed and non-indexed event parameters must be treated separately
+        //indexed parameters are stored in the 'topics' array
+        //non-indexed parameters are stored in the 'data' value
+        $eventParameterNames = [];
+        $eventParameterTypes = [];
+        $eventIndexedParameterNames = [];
+        $eventIndexedParameterTypes = [];
+
+        foreach ($this->events[$eventName]['inputs'] as $input) {
+            if ($input['indexed']) {
+                $eventIndexedParameterNames[] = $input['name'];
+                $eventIndexedParameterTypes[] = $input['type'];
+            } else {
+                $eventParameterNames[] = $input['name'];
+                $eventParameterTypes[] = $input['type'];
+            }
+        }
+
+        $numEventIndexedParameterNames = count($eventIndexedParameterNames);
+
+        //filter through log data to find any logs which match this event (topic) from
+        //this contract, between these specified blocks (defaulting to the latest block only)
+        $this->eth->getLogs([
+            'fromBlock' => (is_int($fromBlock)) ? '0x' . dechex($fromBlock) : $fromBlock,
+            'toBlock' => (is_int($toBlock)) ? '0x' . dechex($toBlock) : $toBlock,
+            'topics' => [$this->ethabi->encodeEventSignature($this->events[$eventName])],
+            'address' => $this->toAddress
+        ],
+            function ($error, $result) use (&$eventLogData, $eventParameterTypes, $eventParameterNames, $eventIndexedParameterTypes, $eventIndexedParameterNames,$numEventIndexedParameterNames) {
+                if ($error !== null) {
+                    throw new RuntimeException($error->getMessage());
+                }
+
+                foreach ($result as $object) {
+                    //decode the data from the log into the expected formats, with its corresponding named key
+                    $decodedData = array_combine($eventParameterNames, $this->ethabi->decodeParameters($eventParameterTypes, $object->data));
+
+                    //decode the indexed parameter data
+                    for ($i = 0; $i < $numEventIndexedParameterNames; $i++) {
+                        //topics[0] is the event signature, so we start from $i + 1 for the indexed parameter data
+                        $decodedData[$eventIndexedParameterNames[$i]] = $this->ethabi->decodeParameters([$eventIndexedParameterTypes[$i]], $object->topics[$i + 1])[0];
+                    }
+
+                    //include block metadata for context, along with event data
+                    $eventLogData[] = [
+                        'transactionHash' => $object->transactionHash,
+                        'blockHash' => $object->blockHash,
+                        'blockNumber' => hexdec($object->blockNumber),
+                        'data' => $decodedData
+                    ];
+                }
+            });
+
+        return $eventLogData;
     }
 }